+++ packages/docusaurus-search-local/src/client/theme/SearchBar/generatedWrapper.ts
@@ -7,6 +7,7 @@ export const docsBasePath: string = data.docsBasePath;
 export const tokenize: (input: string) => string[] = data.tokenize;
 export const titleBoost: number = data.titleBoost;
 export const contentBoost: number = data.contentBoost;
+export const parentCategoriesBoost: number = data.parentCategoriesBoost;
 export const indexDocSidebarParentCategories: boolean =
   data.indexDocSidebarParentCategories;
 export const mylunr: typeof lunr = data.mylunr;