+++ packages/docusaurus-search-local/src/server/index.ts
@@ -35,10 +35,11 @@ type MyOptions = {
   indexDocSidebarParentCategories: number;
   lunr: {
     tokenizerSeparator?: string;
-    k1?: number;
-    b?: number;
-    titleBoost?: number;
-    contentBoost?: number;
+    k1: number;
+    b: number;
+    titleBoost: number;
+    contentBoost: number;
+    parentCategoriesBoost: number;
   };
   style?: "none";
 };
@@ -94,9 +95,10 @@ const optionsSchema = Joi.object({
   lunr: Joi.object({
     tokenizerSeparator: Joi.object().regex(),
     b: Joi.number().min(0).max(1).default(0.75),
-    k1: Joi.number().min(0).max(5).default(1.2), // TODO not sure 5 is a good max
-    contentBoost: Joi.number().min(0).default(1),
+    k1: Joi.number().min(0).default(1.2),
     titleBoost: Joi.number().min(0).default(5),
+    contentBoost: Joi.number().min(0).default(1),
+    parentCategoriesBoost: Joi.number().min(0).default(2),
   }).default(),
 });
 
@@ -120,16 +122,9 @@ export default function cmfcmfDocusaurusSearchLocal(
       b,
       titleBoost,
       contentBoost,
+      parentCategoriesBoost,
     },
   } = options;
-
-  logger.info(`
---- lunr options ---
-tokenizerSeparator: ${lunrTokenizerSeparator || "default"}
-k1: ${k1}
-b: ${b}
-titleBoost: ${titleBoost}
-contentBoost: ${contentBoost}`);
 
   if (lunrTokenizerSeparator) {
     // @ts-expect-error
@@ -250,6 +245,7 @@ export const tokenize = (input) => lunr.tokenizer(input)
   generated += `export const mylunr = lunr;\n`;
   generated += `export const titleBoost = ${titleBoost};\n`;
   generated += `export const contentBoost = ${contentBoost};\n`;
+  generated += `export const parentCategoriesBoost = ${parentCategoriesBoost};\n`;
   generated += `export const docsBasePath = ${JSON.stringify(docsBasePath)};\n`;
   generated += `export const blogBasePath = ${JSON.stringify(blogBasePath)};\n`;
   generated += `export const indexDocSidebarParentCategories = ${JSON.stringify(
@@ -372,13 +368,13 @@ export const tokenize = (input) => lunr.tokenizer(input)
             this.use(lunr[language]);
           }
         }
+
+        this.k1(k1);
+        this.b(b);
+
         this.ref("id");
         this.field("title");
         this.field("content");
-        logger.info(`Using 'k1' of ${k1}.`);
-        this.k1(k1!); // controls how quickly the boost given by a common word reaches saturation
-        logger.info(`Using 'b' of ${b}.`);
-        this.b(b!); // controls the importance given to the length of a document and its fields.
 
         if (useDocVersioning) {
           this.field("version");