+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - FEAT: Add option to specify the path to the docs directory (#78)
+- FEAT: Make boosting behavior of title vs content vs parent categories configurable (#67)
 - BREAKING: Now requires Docusaurus v2.0.0-beta4 or later
 - BREAKING: Now requires at least Node.js 14
 - CLEANUP: Reorganize repository layout, use yarn workspaces